+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/internal/model/DbTransactionOverview.kt
@@ -23,7 +23,8 @@ internal data class DbTransactionOverview internal constructor(
     val memoCount: Int,
     val blockTimeEpochSeconds: Long?,
     val isShielding: Boolean,
-    val isExpiredUnmined: Boolean?
+    val isExpiredUnmined: Boolean?,
+    val poolCrossingValue: Zatoshi?
 ) {
     override fun toString() = "DbTransactionOverview"